requirements.txt
asgiref==3.11.0
Django==5.2.10
django-ckeditor==6.7.3
django-jazzmin==3.0.1
django-js-asset==3.1.2
pillow==12.1.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
django-image-cropping==1.7
easy-thumbnails==2.10.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
django-storages[google]==1.14.4
google-cloud-storage==2.19.0
dj-database-url==2.3.0
xhtml2pdf==0.2.17
django-redis==5.4.0
redis==5.0.1
django-debug-toolbar==4.4.6
django-ipware==7.0.1